Giodev

+++ api/schemas/chat.py
@@ -26,7 +26,7 @@ class RegionalScoreContext(BaseModel):
     ppst_research_based_practice: float
     ppst_assessment_literacy: float
     ppst_professional_development: float
-    demand_signal_count: float
+    demand_signal_count: int
     critical_pings: list[CriticalPing] | None = None
